py/konsulko-labgrid: Add sample rpi_arm64 / rpi_4 boardenv file
To match up with the sample conf file under bin/konsulko-labgrid, add a
boardenv file for the rpi_4 and rpi_arm64 variants (cannot be shared
with rpi_4_32b as we have aarch64 specific example test files). This
shows how to enable network based tests, including OS boot on the
platform.

Signed-off-by: Tom Rini <trini@konsulko.com>