labgrid: ti-omap: Add writer using write-files
While TI OMAP platforms can find the "MLO" in a number of ways on an SD card,
the most common method is to have the first partition be a FAT filesystem and
for MLO and then u-boot.img to be written there. Add a script to do this via
the write-files labgrid-client sub-command.

Signed-off-by: Tom Rini <trini@konsulko.com>